pyproject.toml
[tool.poetry]
name = "venv"
version = "0.1.0"
description = ""
authors = ["HiRurgLipay <sl38lipay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
requests = "^2.31.0"
pip = "^24.0"
install = "^1.3.5"
djangorestframework = "^3.14.0"
redis = "^5.0.1"
celery = "^5.3.6"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
